backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
alembic==1.12.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File upload
python-multipart==0.0.6

# Redis (optional - can skip if issues)
redis==5.0.1

# Data processing and forecasting
pandas==2.1.3
numpy==1.26.2
prophet==1.1.5

# Testing and data generation
faker==20.1.0
pytest==7.4.3
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pytz==2023.3.post1

# Logging
loguru==0.7.2

#OPTIONAL packages (can skip for MVP)::
# Can comment out or remove if install fails:
# hiredis==2.2.3          # Redis optimization (optional)
# openpyxl==3.1.2         # Excel export (not implemented yet)
# xlsxwriter==3.1.9       # Excel export (not implemented yet)
# email-validator==2.1.0  # Email validation (not using auth)
# python-jose[cryptography]==3.3.0  # JWT auth (skipping for demo)
# passlib[bcrypt]==1.7.4  # Password hashing (skipping for demo)